Speed up domain import by computing sensitive domains lazily
Since our users may not actually visit most of the domains we know
about, we can wait until a user actually visits a domain before we
calculate the sensitivity for it.

I also simplified DomainClassifier#isSensitive to always give either a
true or false value for the domain, since the client wants a
true/false value.