Refactored FortyTwoController to allow setting a parser explicitly.
This affected many other files, and changes testing slightly. When merged, we need to make sure all clients don't have an issue with this (since we're not accepting AnyContent anymore). However, as explained below, this should be resilient.

In Play!, `parse.json` requires that the client's Content Type be set correctly, and the body to be valid Json. `parse.tolerantJson` ignores the Content Type, but still requires the body to be valid Json. We have several routes that we want to use AuthenticatedJsonAction because they return Json, but they don't actually use a body. So, `tolerantJson` fails with an empty body. This implementation includes a Json parser called `actuallyTolerantJsonForReals`, which allows an empty body, and sets the body to JsNull.