hack to prevent flood of saveNormalizedUri calls to shoebox
This adds a (machine specific) lock to scraper so that no two calls to
the endpoint so that no to calls can be in progress at the same time.
This is only a bandaid so that when there are a lot of calls it will
starve scraper threads instead of shoebox which is more central to the
system. Real solution, which will take more time:
A) Make all these inter service calls non-blocking/responsive
and
B) Make shoebox throttle the incoming calls through an actor as
necessary
and
C) stop using saveNormalizedUri in favor of updateNormalizedUri