Random Idea: Reactive "Lock"
`withLock`/`withLockFuture` behaves just like a regular `synchronized`
in that it is guaranteed that no two thread will be executing the
enclosing block at the same time (although the `ReactiveLock` also
works for async blocks, which a regular `synchronized` does not) except
that if an execution is in progress any further incoming threads will
not block but rather just get a future back immediately which will only
start executing when it is its turn.
Internally behaves somewhat similar to an actor in that there is a task
queue that is worked through sequentially, however full type safety is
preserved, there is simple queue monitoring and it also supports async
blocks for which you need separate (albeit very easy, since single
threaded) locking inside of an actor.
Of course this is more expensive at point of use than both actors and
regular locks, but it should make is much easier to sequentialize
existing code when needed without a bigger refactor into an actor in
cases where all you want is sequentialization of a particular call or
set of calls.