1 retry in the sequence number assigner
I think the raw_seed_item table in curator is currently the most
insert/save heavy table in the system. We’ve been seeing the occasional
lock wait timeout exceeded error for the sequence number assigner there
when big batch updates/inserts are going on. I think it’s likely only a
matter of time before this happens elsewhere as well, this I would like
to add one retry to the logic.
Thoughts?