email template placeholders can be applied to the "subject" and "fromName" parameters
It's common that we want an email's fromName to be "<first> <last> (via Kifi)".  This creates a light-weight pattern to default to this fromName with only a user id.

- `EmailToSend.fromName` is now `Option[Either[Id[User], String]]`.  instead of Option[String].

If Left(id) is used, it will be preprocessed as "<% fullName(id) %> (via Kifi)" and then feed through the template processor
`Left(id)` is just shorthand for `Right(fullName(id) + " (via Kifi)")`

If `Right(str)` is used, it will still go through the template processor and any tags will be replaced

It was also common to put first and last names in the subject line. Instead of duplicating this, I also made this parameter work with placeholder.

Allowing these fields