basic libraries info in search analytics (kifi.com)
this stuff is going to need some more love soon. Click tracking seems
to have been broken since we switched out the search endpoint, due to a
piece of data (matches) no longer being available. I’ve fixed the
exception, but we need to figure out if we still need that data and if
so get it back into the api.
Also, if a user completely eases the query that does *not* cause a
‘searched’ event. I’m not sure that is what we want.