Cache IPs and notifications for new clusters
Added a cache to the userIpAddressRepo and logic that prevents a user
from being logged if they are still in the cache. Also, upon adding
a user we check to see if they share an ip with any other people. If
this is a new discovery, we report it to the Slack channel #userclusters