Proto Organizations
A proto-Organization is a placeholder organization that holds all the
information necessary to create a real organization (e.g., the owner, name,
and a list of members to invite). There are two db tables for proto orgs,
one for the proto-orgs themselves and another for all the members associated
with them.