Orgs can have proto members. No more proto-orgs.
Instead of a proto-org, we just create a secret organization with a temp
owner and add "proto-members" to that org. We can eventually invite the
proto members to the org and they'll get an actual invite. When we
want to "unveil" the org, we transfer ownership from the temp owner to a
real Kifi user, unhide the org, and invite all the proto-members.