Org avatar commander no longer reuses images
Every time a user uploads an org avatar, it gets processed and uploaded to S3.
It is unlikely that we will be able to reuse anything from the image store
given how many degrees of freedom a cropped scaled image has.

Also changed all the subclasses of ProcessImageRequest into case classes
so we can actually match on them.