Removed all concept of disjoint libraries
Disjoint libraries were the system-generated libraries. Those two exhibited
different behavior, in that they could not share any URIs. If a URI was
interned in one lib, it would be removed from the other. This should
no longer be the case.

Keep dropped the `inDisjointLib` field (this needs to be applied to
the DB!) and Library dropped the `isDisjoint` method.