Changed the way we generate org avatar image paths
Now they are much more descriptive, so they are guaranteed to be unique
for any image.

Previously we used `imagehash_widthxheight_cs.png`. It is possible to have
many different images of the same dimensions that come from the same source,
so this led to collisions in the CDN.

Now we use `imagehas_sxs+x+y->SxS_cs.png` where `s` is the crop size
and `S` is the final scaled size (`S \in {100, 200}`).