Made convenience classes for dealing with billing
Currency represents currency. Under the hood it stores things in `cents: Int`.

BillingCycle represents a time period. Under the hood it is using `months: Int`.

BillingRate represents a currency / time. Under the hood it is storing
just the `cost: Currency`, and it has a fixed time period that it uses.
You can construct a BillingRate by dividing a Currency by a BillingCycle,
and it will throw an exception if they are not evenly divisible. I did
that because it is the only way to ensure that
  1. We are billing the same amount of money every month
  2. We are billing the correct amount of money over the entire billing cycle

If we are willing to relax either of those constraints then we can do other
things (maybe charge a different amount on the last month of a billing cycle).

Now that I'm writing this out it seems more reasonable to only charge once per
billing cycle. I think that might even be what billing cycle actually means...