Got Keep.entitiesHash to work
I had a really obnoxious bug where I was serializing the Keep into Json
incorrectly, so the Keep cache was getting garbage data (but the DB was fine).
When keepRepo.get was called it was automatically looking in the cache and
pulling out the garbage data. Tough to track down.