Fix inserts into notification table having a current timestamp
In mysql, specifying 'default null' means that the column becomes a
null column. Inserting null into this column will keep it as null. In
h2, however, only specifying 'default null' means that inserts to the
column will be null, and then become the current timestamp due to sql
behavior. This makes sure that the column is specified as nullablle on
both systems.