Added a General library to orgs
It is a system library. It is created automatically when an org is created. Org
members are automatically added to the library when they join the org, and they
are removed when they leave or are kicked out.

The library cannot be modified or deleted. It automatically transfers ownership
when the org owner changes. It is deleted when the org is deleted.