Fix org checker (session-in-session issues)
Library modify/delete is asynchronous, so it opens its own session. This is
an issue because the org checker was doing all of its modification in a single
monster read/write session. The fix involved splitting the single monster session
into multiple sessions.

This would be super expensive, so we first check all the orgs in a single
read session, pick out just the ones that are broken, then fix each broken
one independently. We block the thread to wait for the async library fixes
to complete.