Factored out credit reward serialization
CreditRewardInfoCommander generates the DescriptionElements for credit rewards
now (the ActivityLogCommander defers to it).

The endpoint /site/organizations/:id/rewards serializes a CreditRewardsView,
which presents all of an organization's rewards (both applied and unearned),
along with descriptions of them. For applied rewards, this will look exactly
like the activity log. For unearned rewards, this contains information about
how to go about earning them (and how much will be earned upon completion).