Made a bunch of library code a bit safer
We shouldn't do things with inactive libraries. Somehow these ids are leaking
out, and there is code that is trying to use them. I made most of this code
more robust: where possible, it assumes that some of the library ids it has
could be invalid, and it does a `libraryIds.flatMap(libraryRepo.getActive)`
thing that drops the inactive ids.