Refactor SlackFail
SlackFail is the most general form of failure involving Slack. It can be
anything from "we do not have any way on file to actually send this message"
to "we tried this and Slack gave us back an error message".

SlackAPIErrorResponse is a wrapper around things that Slack returns. These
come directly from Slack. The best way to match them is to use the
SlackErrorCode.unapply method and directly check the error code.