Invite kifi-bot to channels when it wants to talk
When SlackClientWrapper.sendToSlackViaBot is called, we look for a bot.
If we find a bot, we try and send a message.

The new logic is that IF that push fails because Slack says that kifi-bot isn't
in the channel, we try and find a user from that team that can invite the little
guy into the channel. If we succeed then we try again.