Fix unread count > 0 with no displayable notifs
This issue happened when a user thread was marked as unread, but the
message thread had no messages (and was thus ignored for display purposes).

The fix is just to have user threads with unread=false for the creator of the
message thread, by default. They are set to unread=true when a message is sent