Introduce KeepQueryCommander
This serves a similar role to `LibraryQueryCommander`, namely to provide
a uniform API for user-defined ordering of keeps. Right now the
only area that it can serve is paging through keeps in a single library.

The default behavior is to order by `keep.lastActivityAt` in descending order.
There is currently no way to alter this behavior, but soon I will introduce
an endpoint to let users set their default behavior, and after that I will
work on allowing the frontend to send down custom orderings