Add BasicKeepEventId to all events, split out note
All events now have an id. For "implicit" events (ones that are not explicitly
persisted in the keep_event table, i.e., initial/note events) there is a
pseudo-id of the form "init_<keep-id>" or "note_<keep-id>". This is simpler
to the frontend and more uniform on the backend.

The note event is always paired with the initial keep event. This means that
sometimes we will send down more than the requested number of events. Talked to
Carlos and I don't think this will be an issue.